DVD: Try to better generate missing stop packets
We now handle multiple open start packets per caller id and close them all, when a new (set of) call(s) start. This is because DMR likes to transmit the same call on more than one time slot creating multiple concurrent calls that need to be closed when the stop packets are lost.